+++ lib/pages/challenge/challenge.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:ninety9names/pages/names/details.dart';
-import 'package:ninety9names/pages/names/model.dart';
+import 'package:ninety9names/repo/name.dart';
 import 'package:http/http.dart' as http;
 
 class Challenge extends StatefulWidget {
