+++ lib/repo/name.dart
@@ -1,0 +1,10 @@
+class Name {
+  final String? id;
+  final String? arabic;
+  final String? transliteration;
+  final String? meaning;
+  final String? explanation;
+
+  Name(this.id, this.arabic, this.transliteration, this.meaning,
+      this.explanation);
+}
